pyproject.toml
[project]
name = "mausoleo"
version = "0.1.0"
description = "Hierarchical knowledge index for historical newspaper archives"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.28.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "clickhouse-connect>=0.7.0",
    "python-multipart>=0.0.24",
]

[project.optional-dependencies]
ocr = [
    "ray[data]>=2.9.0",
    "Pillow>=10.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.25.0",
    "torch>=2.0.0",
    "qwen-vl-utils>=0.0.8",
]
ocr-vllm = [
    "vllm>=0.4.0",
]
ocr-yolo = [
    "ultralytics>=8.0.0",
]
ocr-surya = [
    "surya-ocr>=0.6.0",
]
index = [
    "ray[data]>=2.9.0",
    "vllm>=0.4.0",
]
eval = [
    "jiwer>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pyright>=1.1.0",
    "ruff>=0.3.0",
    "ty>=0.0.1a1",
    "ray[data]>=2.9.0",
    "jiwer>=3.0.0",
]

[project.scripts]
mausoleo = "mausoleo.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mausoleo"]

[tool.ruff]
line-length = 144
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
ignore = ["E402", "E702", "E731", "UP008", "E712"]
extend-select = ["UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportMissingImports = "warning"
reportUnknownMemberType = "warning"
reportUnknownVariableType = "warning"
reportUnknownParameterType = "warning"
reportUnknownArgumentType = "warning"
